lowering: split programLowering data/finalization helpers
