+++ src/lowering/programLowering.ts
@@ -31,17 +31,18 @@ import type {
 } from '../formats/types.js';
 import type { CompileEnv } from '../semantics/env.js';
 import type { FunctionLoweringContext } from './functionLowering.js';
-import type { NamedSectionContributionSink, StartupInitAction } from './sectionContributions.js';
+import type { NamedSectionContributionSink } from './sectionContributions.js';
 import type {
   Callable,
   PendingSymbol,
   SectionKind,
 } from './loweringTypes.js';
 import type { AggregateType, ScalarKind } from './typeResolution.js';
+import { lowerDataBlock } from './programLoweringData.js';
 
 // Program lowering owns module-wide declaration traversal and the final
 // emission/fixup passes after all symbols and section bases are known.
-type Context = Omit<FunctionLoweringContext, 'item'> & {
+export type Context = Omit<FunctionLoweringContext, 'item'> & {
   program: ProgramNode;
   includeDirs: string[];
   localCallablesByFile: Map<string, Map<string, Callable>>;
@@ -97,7 +98,7 @@ type Context = Omit<FunctionLoweringContext, 'item'> & {
   withNamedSectionSink: <T>(sink: NamedSectionContributionSink, fn: () => T) => T;
 };
 
-type FinalizationContext = {
+export type FinalizationContext = {
   diagnostics: Diagnostic[];
   diag: (diagnostics: Diagnostic[], file: string, message: string) => void;
   primaryFile: string;
@@ -637,433 +638,4 @@ export function lowerProgramDeclarations(ctx: Context): void {
   }
 }
 
-function lowerDataBlock(
-  ctx: Context,
-  dataBlock: DataBlockNode,
-  target: {
-    section: 'code' | 'data';
-    bytes: Map<number, number>;
-    offsetRef: { current: number };
-    pending: PendingSymbol[];
-    startupInitActions?: StartupInitAction[];
-  } = {
-    section: 'data',
-    bytes: ctx.dataBytes,
-    offsetRef: ctx.dataOffsetRef,
-    pending: ctx.pending,
-  },
-): void {
-  for (const decl of dataBlock.decls) {
-    const okToDeclareSymbol = !ctx.taken.has(decl.name);
-    if (!okToDeclareSymbol) {
-      ctx.diag(ctx.diagnostics, decl.span.file, `Duplicate symbol name "${decl.name}".`);
-    } else {
-      ctx.taken.add(decl.name);
-      target.pending.push({
-        kind: 'data',
-        name: decl.name,
-        section: target.section,
-        offset: target.offsetRef.current,
-        file: decl.span.file,
-        line: decl.span.start.line,
-        scope: 'global',
-      });
-    }
-
-    const type = decl.typeExpr;
-    const init = decl.initializer;
-
-    const emitByte = (b: number) => {
-      target.bytes.set(target.offsetRef.current, b & 0xff);
-      target.offsetRef.current++;
-    };
-    const emitWord = (w: number) => {
-      emitByte(w & 0xff);
-      emitByte((w >> 8) & 0xff);
-    };
-    const nextPow2 = (value: number): number => {
-      if (value <= 1) return value;
-      let pow = 1;
-      while (pow < value) pow <<= 1;
-      return pow;
-    };
-    const recordStartupInit = (
-      kind: 'copy' | 'zero',
-      offset: number,
-      length: number,
-    ) => {
-      if (!target.startupInitActions) return;
-      if (length <= 0) return;
-      target.startupInitActions.push({
-        kind,
-        offset,
-        length,
-      });
-    };
-
-    const recordType = ctx.resolveAggregateType(type);
-    if (recordType?.kind === 'record') {
-      if (init.kind === 'InitZero') {
-        const zeroStart = target.offsetRef.current;
-        const storageBytes = ctx.sizeOfTypeExpr(type, ctx.env, ctx.diagnostics);
-        if (storageBytes === undefined) continue;
-        for (let pad = 0; pad < storageBytes; pad++) emitByte(0);
-        recordStartupInit('zero', zeroStart, storageBytes);
-        continue;
-      }
-      if (init.kind === 'InitString') {
-        ctx.diag(ctx.diagnostics, decl.span.file, `Record initializer for "${decl.name}" must use aggregate form.`);
-        continue;
-      }
-
-      const valuesByField = new Map<string, ImmExprNode>();
-      let recordInitFailed = false;
-      if (init.kind === 'InitRecordNamed') {
-        for (const fieldInit of init.fields) {
-          const field = recordType.fields.find((f) => f.name === fieldInit.name);
-          if (!field) {
-            ctx.diag(ctx.diagnostics, decl.span.file, `Unknown record field "${fieldInit.name}" in initializer for "${decl.name}".`);
-            recordInitFailed = true;
-            continue;
-          }
-          if (valuesByField.has(field.name)) {
-            ctx.diag(ctx.diagnostics, decl.span.file, `Duplicate record field "${field.name}" in initializer for "${decl.name}".`);
-            recordInitFailed = true;
-            continue;
-          }
-          valuesByField.set(field.name, fieldInit.value);
-        }
-        for (const field of recordType.fields) {
-          if (valuesByField.has(field.name)) continue;
-          ctx.diag(ctx.diagnostics, decl.span.file, `Missing record field "${field.name}" in initializer for "${decl.name}".`);
-          recordInitFailed = true;
-        }
-      } else {
-        if (init.elements.length !== recordType.fields.length) {
-          ctx.diag(ctx.diagnostics, decl.span.file, `Record initializer field count mismatch for "${decl.name}".`);
-          continue;
-        }
-        for (let index = 0; index < recordType.fields.length; index++) {
-          const field = recordType.fields[index]!;
-          const element = init.elements[index]!;
-          valuesByField.set(field.name, element);
-        }
-      }
-      if (recordInitFailed) continue;
-
-      const encodedFields: Array<{ width: 1 | 2; value: number }> = [];
-      for (const field of recordType.fields) {
-        const fieldValueExpr = valuesByField.get(field.name);
-        if (!fieldValueExpr) continue;
-        const scalar = ctx.resolveScalarKind(field.typeExpr);
-        if (!scalar) {
-          ctx.diag(ctx.diagnostics, decl.span.file, `Unsupported record field type "${field.name}" in initializer for "${decl.name}" (expected byte/word/addr/ptr).`);
-          recordInitFailed = true;
-          continue;
-        }
-        const value = ctx.evalImmExpr(fieldValueExpr, ctx.env, ctx.diagnostics);
-        if (value === undefined) {
-          ctx.diag(ctx.diagnostics, decl.span.file, `Failed to evaluate data initializer for "${decl.name}".`);
-          recordInitFailed = true;
-          continue;
-        }
-        encodedFields.push({ width: scalar === 'byte' ? 1 : 2, value });
-      }
-      if (recordInitFailed) continue;
-
-      let emitted = 0;
-      const copyStart = target.offsetRef.current;
-      for (const encoded of encodedFields) {
-        if (encoded.width === 1) {
-          emitByte(encoded.value);
-          emitted += 1;
-        } else {
-          emitWord(encoded.value);
-          emitted += 2;
-        }
-      }
-      const storageBytes = ctx.sizeOfTypeExpr(type, ctx.env, ctx.diagnostics);
-      if (storageBytes === undefined) continue;
-      recordStartupInit('copy', copyStart, emitted);
-      const zeroStart = target.offsetRef.current;
-      for (let pad = emitted; pad < storageBytes; pad++) emitByte(0);
-      recordStartupInit('zero', zeroStart, storageBytes - emitted);
-      continue;
-    }
-
-    if (init.kind === 'InitRecordNamed') {
-      ctx.diag(ctx.diagnostics, decl.span.file, `Named-field aggregate initializer requires a record type for "${decl.name}".`);
-      continue;
-    }
-
-    if (init.kind === 'InitZero') {
-      const zeroStart = target.offsetRef.current;
-      const storageBytes = ctx.sizeOfTypeExpr(type, ctx.env, ctx.diagnostics);
-      if (storageBytes === undefined) continue;
-      for (let pad = 0; pad < storageBytes; pad++) emitByte(0);
-      recordStartupInit('zero', zeroStart, storageBytes);
-      continue;
-    }
-
-    const elementScalar = type.kind === 'ArrayType' ? ctx.resolveScalarKind(type.element) : ctx.resolveScalarKind(type);
-    const elementSize = elementScalar === 'word' || elementScalar === 'addr' ? 2 : elementScalar === 'byte' ? 1 : undefined;
-    if (!elementSize) {
-      ctx.diag(ctx.diagnostics, decl.span.file, `Unsupported data type for "${decl.name}" (expected byte/word/addr/ptr or fixed-length arrays of those).`);
-      continue;
-    }
-
-    const declaredLength = type.kind === 'ArrayType' ? type.length : 1;
-    let actualLength = declaredLength ?? 0;
-
-    if (init.kind === 'InitString') {
-      if (elementSize !== 1) {
-        ctx.diag(ctx.diagnostics, decl.span.file, `String initializer requires byte element type for "${decl.name}".`);
-        continue;
-      }
-      if (declaredLength !== undefined && init.value.length !== declaredLength) {
-        ctx.diag(ctx.diagnostics, decl.span.file, `String length mismatch for "${decl.name}".`);
-        continue;
-      }
-      const copyStart = target.offsetRef.current;
-      for (let idx = 0; idx < init.value.length; idx++) emitByte(init.value.charCodeAt(idx));
-      actualLength = init.value.length;
-      recordStartupInit('copy', copyStart, actualLength);
-      if (type.kind === 'ArrayType') {
-        const emittedBytes = actualLength * elementSize;
-        const storageBytes = nextPow2(emittedBytes);
-        const zeroStart = target.offsetRef.current;
-        for (let pad = emittedBytes; pad < storageBytes; pad++) emitByte(0);
-        recordStartupInit('zero', zeroStart, storageBytes - emittedBytes);
-      }
-      continue;
-    }
-
-    const values: number[] = [];
-    for (const e of init.elements) {
-      const v = ctx.evalImmExpr(e, ctx.env, ctx.diagnostics);
-      if (v === undefined) {
-        ctx.diag(ctx.diagnostics, decl.span.file, `Failed to evaluate data initializer for "${decl.name}".`);
-        break;
-      }
-      values.push(v);
-    }
-
-    if (declaredLength !== undefined && values.length !== declaredLength) {
-      ctx.diag(ctx.diagnostics, decl.span.file, `Initializer length mismatch for "${decl.name}".`);
-      continue;
-    }
-
-    const copyStart = target.offsetRef.current;
-    for (const v of values) {
-      if (elementSize === 1) emitByte(v);
-      else emitWord(v);
-    }
-    actualLength = type.kind === 'ArrayType' ? values.length : 1;
-    recordStartupInit('copy', copyStart, actualLength * elementSize);
-    if (type.kind === 'ArrayType') {
-      const emittedBytes = actualLength * elementSize;
-      const storageBytes = nextPow2(emittedBytes);
-      const zeroStart = target.offsetRef.current;
-      for (let pad = emittedBytes; pad < storageBytes; pad++) emitByte(0);
-      recordStartupInit('zero', zeroStart, storageBytes - emittedBytes);
-    }
-  }
-}
-
-export function finalizeProgramEmission(ctx: FinalizationContext): {
-  codeBase: number;
-  dataBase: number;
-  varBase: number;
-  codeOk: boolean;
-  dataOk: boolean;
-  varOk: boolean;
-  writtenRange: AddressRange;
-  sourceSegments: EmittedSourceSegment[];
-  asmTrace: EmittedAsmTraceEntry[];
-} {
-  const evalBase = (kind: SectionKind): number | undefined => {
-    const at = ctx.baseExprs[kind];
-    if (!at) return undefined;
-    const value = ctx.evalImmExpr(at, ctx.env, ctx.diagnostics);
-    if (value === undefined) {
-      ctx.diag(ctx.diagnostics, at.span.file, `Failed to evaluate section "${kind}" base address.`);
-      return undefined;
-    }
-    if (value < 0 || value > 0xffff) {
-      ctx.diag(ctx.diagnostics, at.span.file, `Section "${kind}" base address out of range (0..65535).`);
-      return undefined;
-    }
-    return value;
-  };
-
-  const explicitCodeBase = evalBase('code');
-  const explicitDataBase = evalBase('data');
-  const explicitVarBase = evalBase('var');
-  const codeOk = explicitCodeBase !== undefined || !ctx.baseExprs.code;
-  const codeBase = explicitCodeBase ?? (ctx.defaultCodeBase ?? 0);
-  const dataBase =
-    explicitDataBase ??
-    (codeOk
-      ? ctx.alignTo(codeBase + ctx.codeOffset, 2)
-      : (ctx.diag(
-          ctx.diagnostics,
-          ctx.primaryFile,
-          `Cannot compute default data base address because code base address is invalid.`,
-        ),
-        0));
-  const dataOk = explicitDataBase !== undefined || (ctx.baseExprs.data === undefined && codeOk);
-  const varBase =
-    explicitVarBase ??
-    (dataOk
-      ? ctx.alignTo(dataBase + ctx.dataOffset, 2)
-      : (ctx.diag(
-          ctx.diagnostics,
-          ctx.primaryFile,
-          `Cannot compute default var base address because data base address is invalid.`,
-        ),
-        0));
-  const varOk = explicitVarBase !== undefined || (ctx.baseExprs.var === undefined && dataOk);
-
-  const addrByNameLower = new Map<string, number>();
-  for (const ps of ctx.pending) {
-    const base = ps.section === 'code' ? codeBase : ps.section === 'data' ? dataBase : varBase;
-    const ok = ps.section === 'code' ? codeOk : ps.section === 'data' ? dataOk : varOk;
-    if (!ok) continue;
-    addrByNameLower.set(ps.name.toLowerCase(), base + ps.offset);
-  }
-  for (const sym of ctx.symbols) {
-    if (sym.kind === 'constant') continue;
-    addrByNameLower.set(sym.name.toLowerCase(), sym.address);
-  }
-  for (const sym of ctx.absoluteSymbols) {
-    if (sym.kind === 'constant' || sym.address === undefined) continue;
-    addrByNameLower.set(sym.name.toLowerCase(), sym.address);
-  }
-  for (const ex of ctx.deferredExterns) {
-    const base = addrByNameLower.get(ex.baseLower);
-    if (base === undefined) {
-      ctx.diag(
-        ctx.diagnostics,
-        ex.file,
-        `Failed to resolve extern base symbol "${ex.baseLower}" for "${ex.name}".`,
-      );
-      continue;
-    }
-    const addr = base + ex.addend;
-    if (addr < 0 || addr > 0xffff) {
-      ctx.diag(
-        ctx.diagnostics,
-        ex.file,
-        `extern func "${ex.name}" resolved address out of range (0..65535).`,
-      );
-      continue;
-    }
-    addrByNameLower.set(ex.name.toLowerCase(), addr);
-    ctx.symbols.push({
-      kind: 'label',
-      name: ex.name,
-      address: addr,
-      file: ex.file,
-      line: ex.line,
-      scope: 'global',
-    });
-  }
-
-  // Absolute fixups are resolved in a dedicated second pass after all labels,
-  // data bases, and deferred extern addresses are known.
-  for (const fx of ctx.fixups) {
-    const base = addrByNameLower.get(fx.baseLower);
-    const addr = base === undefined ? undefined : base + fx.addend;
-    if (addr === undefined) {
-      ctx.diag(ctx.diagnostics, fx.file, `Unresolved symbol "${fx.baseLower}" in 16-bit fixup.`);
-      continue;
-    }
-    if (addr < 0 || addr > 0xffff) {
-      ctx.diag(
-        ctx.diagnostics,
-        fx.file,
-        `16-bit fixup address out of range for "${fx.baseLower}" with addend ${fx.addend}: ${addr}.`,
-      );
-      continue;
-    }
-    ctx.codeBytes.set(fx.offset, addr & 0xff);
-    ctx.codeBytes.set(fx.offset + 1, (addr >> 8) & 0xff);
-  }
-
-  // rel8 fixups must also wait until code base + label addresses are finalized
-  // so branch displacements are computed from final absolute origins.
-  for (const fx of ctx.rel8Fixups) {
-    const base = addrByNameLower.get(fx.baseLower);
-    const target = base === undefined ? undefined : base + fx.addend;
-    if (target === undefined) {
-      ctx.diag(
-        ctx.diagnostics,
-        fx.file,
-        `Unresolved symbol "${fx.baseLower}" in rel8 ${fx.mnemonic} fixup.`,
-      );
-      continue;
-    }
-    const origin = codeBase + fx.origin;
-    const disp = target - origin;
-    if (disp < -128 || disp > 127) {
-      ctx.diag(
-        ctx.diagnostics,
-        fx.file,
-        `${fx.mnemonic} target out of range for rel8 branch (${disp}, expected -128..127).`,
-      );
-      continue;
-    }
-    ctx.codeBytes.set(fx.offset, disp & 0xff);
-  }
-
-  for (const [addr, b] of ctx.hexBytes) {
-    if (addr < 0 || addr > 0xffff) {
-      ctx.diag(ctx.diagnostics, ctx.primaryFile, `HEX byte address out of range: ${addr}.`);
-      continue;
-    }
-    if (ctx.bytes.has(addr)) {
-      ctx.diag(ctx.diagnostics, ctx.primaryFile, `HEX data overlaps emitted bytes at address ${addr}.`);
-      continue;
-    }
-    ctx.bytes.set(addr, b);
-  }
-
-  if (codeOk) {
-    ctx.writeSection(codeBase, ctx.codeBytes, ctx.bytes, (message) =>
-      ctx.diag(ctx.diagnostics, ctx.primaryFile, message),
-    );
-  }
-  if (dataOk) {
-    ctx.writeSection(dataBase, ctx.dataBytes, ctx.bytes, (message) =>
-      ctx.diag(ctx.diagnostics, ctx.primaryFile, message),
-    );
-  }
-
-  for (const ps of ctx.pending) {
-    const base = ps.section === 'code' ? codeBase : ps.section === 'data' ? dataBase : varBase;
-    const ok = ps.section === 'code' ? codeOk : ps.section === 'data' ? dataOk : varOk;
-    if (!ok) continue;
-    ctx.symbols.push({
-      kind: ps.kind,
-      name: ps.name,
-      address: base + ps.offset,
-      ...(ps.file !== undefined ? { file: ps.file } : {}),
-      ...(ps.line !== undefined ? { line: ps.line } : {}),
-      ...(ps.scope !== undefined ? { scope: ps.scope } : {}),
-      ...(ps.size !== undefined ? { size: ps.size } : {}),
-    });
-  }
-  ctx.symbols.push(...ctx.absoluteSymbols);
-
-  return {
-    codeBase,
-    dataBase,
-    varBase,
-    codeOk,
-    dataOk,
-    varOk,
-    writtenRange: ctx.computeWrittenRange(ctx.bytes),
-    sourceSegments: codeOk ? ctx.rebaseCodeSourceSegments(codeBase, ctx.codeSourceSegments) : [],
-    asmTrace: codeOk ? ctx.rebaseAsmTrace(codeBase, ctx.codeAsmTrace) : [],
-  };
-}
+export { finalizeProgramEmission } from './programLoweringFinalize.js';
